refactor: rename to createFontFamilyUtils
